LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-flexbox/negative-overflow-002-expected.txt

1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3
1
2
3

PASS .container 1
PASS .container 2
PASS .container 3
PASS .container 4
PASS .container 5
PASS .container 6
PASS .container 7
PASS .container 8
PASS .container 9
PASS .container 10
PASS .container 11
PASS .container 12
FAIL .container 13 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: row;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 355
FAIL .container 14 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 115
FAIL .container 15 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: wrap-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 115
PASS .container 16
PASS .container 17
PASS .container 18
FAIL .container 19 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: column;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 115
FAIL .container 20 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: column wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 355
PASS .container 21
FAIL .container 22 assert_equals:
<div class="container" data-expected-scroll-width="130" data-expected-scroll-height="370" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: column-reverse;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 130 but got 115
FAIL .container 23 assert_equals:
<div class="container" data-expected-scroll-width="370" data-expected-scroll-height="130" style="writing-mode: horizontal-tb; direction: rtl; flex-flow: column-reverse wrap;"><div class="item">1</div><div class="item">2</div><div class="item">3</div></div>
scrollWidth expected 370 but got 355
PASS .container 24
PASS .container 25
PASS .container 26
PASS .container 27
PASS .container 28
PASS .container 29
PASS .container 30
PASS .container 31
PASS .container 32
PASS .container 33
PASS .container 34
PASS .container 35
PASS .container 36
PASS .container 37
PASS .container 38
PASS .container 39
PASS .container 40
PASS .container 41
PASS .container 42
PASS .container 43
PASS .container 44
PASS .container 45
PASS .container 46
PASS .container 47
PASS .container 48
PASS .container 49
PASS .container 50
PASS .container 51
PASS .container 52
PASS .container 53
PASS .container 54
PASS .container 55
PASS .container 56
PASS .container 57
PASS .container 58
PASS .container 59
PASS .container 60
PASS .container 61
PASS .container 62
PASS .container 63
PASS .container 64
PASS .container 65
PASS .container 66
PASS .container 67
PASS .container 68
PASS .container 69
PASS .container 70
PASS .container 71
PASS .container 72
